moved the code from free.c and free.h to malloc.c and malloc.h, updated the makefile
