+++ .gitignore
@@ -42,5 +42,4 @@
 
 # the implementation files, remove this later.
 malloc
-free
 docs
