+++ main.c
@@ -1,0 +1,5 @@
+#include "src/malloc.h"
+
+int main() {
+    return 0;
+}
